# docs/openapi-proxy.yaml
openapi: 3.0.3
info:
  title: ZimaOS Echo - API Proxy Sidecar
  description: |
    API Proxy Sidecar provides intelligent routing, failover, security monitoring,
    and usage statistics for LLM API calls.
  version: 0.10.5
  contact:
    name: IceWhale Tech
    url: https://github.com/IceWhaleTech/ZimaOS-Echo
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:{port}/api/v1/proxy
    description: Local proxy server
    variables:
      port:
        default: "19000"
        description: Dynamically allocated port
...
paths:
...
  /auth/keys:
    get:
      tags:
        - Auth
      summary: List API keys
      description: Returns a list of API keys (masked)
      operationId: getApiKeys
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeysResponse"
    post:
      tags:
        - Auth
      summary: Add API key
      description: Adds a new API key
      operationId: addApiKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyInput"
      responses:
        "200":
          description: Key added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
    delete:
      tags:
        - Auth
      summary: Remove API key
      description: Removes an API key
      operationId: removeApiKey
      parameters:
        - name: key
          in: query
          required: true
          description: The API key to remove
          schema:
            type: string
      responses:
        "200":
          description: Key removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    ApiKey:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
          description: Masked API key
        active:
          type: boolean
    ApiKeysResponse:
      type: object
      properties:
        keys:
          type: array
          items:
            $ref: "#/components/schemas/ApiKey"
    ApiKeyInput:
      type: object
      required:
        - key
      properties:
        key:
          type: string
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - ApiKeyAuth: []